requirements.txt
celery==5.5.2
fastapi==0.115.12
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
redis==5.3.0
uvicorn==0.34.2
